% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{annotate_differences}
\alias{annotate_differences}
\title{Annotate differences with segmentation info}
\usage{
annotate_differences(diff_call, segments, min.overlap.pc = 30)
}
\arguments{
\item{diff_call}{the differences data frame, as returned by call_differences}

\item{segments}{the segments, as output by segment_methylation}

\item{min.overlap.pc}{differences which have no big overlap with any segment
are classified as "other" (default 30 percent)}
}
\description{
Annotate differences with segmentation info
}
