% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_differences}
\alias{call_differences}
\title{call differences}
\usage{
call_differences(data, ret, min.diff = 0.25, fdr.cutoff = 0.05, cov.score=70,
  min.num.cpgs = 4, tol_val = 0.01, ncores = 1)
}
\arguments{
\item{data}{the input data frame}

\item{ret}{the output of difference_detection}

\item{min.diff}{minimum threshold for which a difference should be considered (default 0.25)}

\item{fdr.cutoff}{cutoff for FDR (default 0.05)}

\item{cov.score}{score of coverage for DMR (default 70)}

\item{min.num.cpgs}{Require reported differences to have a certain minimum number of CpGs (default 4)}

\item{tol_val}{fused values below tol_val from each other are considered equal (default 0.01)}

\item{ncores}{number of cores to parallelize on (default=1)}
}
\description{
call differences
}
