% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_differences_singlechr}
\alias{call_differences_singlechr}
\title{call differences}
\usage{
call_differences_singlechr(data, ret, min.delta.diff = 0.25,
  tol_val = 0.01)
}
\description{
single chromosome version
}
