% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide.R
\name{combine_ret}
\alias{combine_ret}
\title{Helper to combine two return values from signal/difference calls}
\usage{
combine_ret(ret1, ret2)
}
\description{
Helper to combine two return values from signal/difference calls
}
