% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide.R
\name{difference_detection_fast}
\alias{difference_detection_fast}
\title{Methylation difference detection using fused lasso, for 2 datasets (each can
have replicates)}
\usage{
difference_detection_fast(data, ref, lambda2 = 50, tol_val = 0.01,
  max_iter = 1, ncores = 1)
}
\arguments{
\item{data}{dataframe containing columns dataset, condition, replicate, chr, pos, Nmethyl and coverage.
dataset, condition and replicate are integer vectors or factors, starting at 1 and increasing without gaps.
chr is the chromosome, as a character or factor. Each chromosome is treated separately.
pos is the genomic coordinate of each CpG, sorted increasingly. Nmethyl is the number
of methylated reads. Ntot is the number of unmethylated reads.}

\item{ref}{the ID of the reference condition}

\item{lambda2}{the initial value of lambda2 (default 50)}

\item{tol_val}{Target precision for the estimated methylation value (default 0.01)}

\item{max_iter}{maximum number of IRLS iterations for the beta binomial model}

\item{ncores}{number of cores to parallelize on (default 1)}
}
\description{
Fast model: normal distribution with identity link and distance-independent modelling
}
