% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_diff.R
\name{difference_detection_fast_singlechr}
\alias{difference_detection_fast_singlechr}
\title{Methylation difference detection using fused lasso, for 2 datasets (each can
have replicates)}
\usage{
difference_detection_fast_singlechr(data, ref, lambda2 = 50,
  tol_val = 0.01, max_iter = 1)
}
\description{
Fast model: normal distribution with identity link and distance-independent modelling
}
\details{
single chromosome
}
