% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLasso.R
\docType{package}
\name{MethyLasso}
\alias{MethyLasso}
\alias{MethyLasso-package}
\title{MethyLasso: a package to detect differences in WGBS data}
\description{
This package performs difference detection on WGBS data.
}
