% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{segment_methylation}
\alias{segment_methylation}
\title{Segment methylation data}
\usage{
segment_methylation(data, ret, ncores = 1, ...)
}
\arguments{
\item{data}{the input data frame}

\item{ret}{the output of signal_detection}

\item{ncores}{number of cores to parallelize on (default=1)}
}
\description{
Segment methylation data
}
