% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide.R
\name{signal_detection}
\alias{signal_detection}
\title{Methylation detection using fused lasso, for 1 dataset or several replicates}
\usage{
signal_detection(data, optimize_lambda2 = F, lambda2 = 50,
  max_lambda2 = 1000, optimize_hyper = F, hyper = 100, nouter = 20,
  bf_per_kb = 50, tol_val = 0.01, max_iter = 30, ncores = 1)
}
\arguments{
\item{data}{dataframe containing columns dataset, condition, replicate, chr, pos, Nmethyl and coverage.
dataset, condition and replicate are integer vectors or factors, starting at 1 and increasing without gaps.
chr is the chromosome, as a character or factor. Each chromosome is treated separately.
pos is the genomic coordinate of each CpG, sorted increasingly. Nmethyl is the number
of methylated reads. Ntot is the number of unmethylated reads.}

\item{optimize_lambda2}{whether to optimize fusion penalty or not (default FALSE)}

\item{lambda2}{the initial value of lambda2 (default 50)}

\item{max_lambda2}{the maximum value of lambda2 (default 1000)}

\item{optimize_hyper}{whether to optimize it (default FALSE)}

\item{hyper}{the value of the hyperparameter (default 100)}

\item{nouter}{the number of iterations to optimize hyper and/or lambda2 (default 20)}

\item{bf_per_kb}{Number of basis functions per kb (default 50)}

\item{tol_val}{Target precision for the estimated methylation value (default 0.01)}

\item{max_iter}{maximum number of IRLS iterations for the beta binomial model}

\item{ncores}{number of cores to parallelize on (default 1)}
}
\description{
Uses beta binomial distribution with logit link
}
